Add files via upload
In 2018, United States President Donald Trump enacted a family separation policy at the US-Mexico border. Under the policy, as immigrant families entered the United States, children were separated from the rest of their families by U.S. Immigration and Customs Enforcement (ICE) and brought to detention facilities around the country.

Upon learning about this policy, the American Civil Liberties Union (ACLU) filed a lawsuit against the administration. The government subsequently provided the ACLU with data related to the immigrant children separated from their families at the border.

As a concerned citizen of the world, you are interested in researching the family separations and want to get a better understanding of how many children were separated and where they were taken. Using the government data provided to the ACLU, you’ll dig into the numbers and gain an understanding of the situation.

Note: A data analysis can only be as good as the data itself. The ACLU has noted concerns about the comprehensiveness and accuracy of the government’s data. Despite this, working with the data available can still yield important insights.

If you get stuck or want to see an experienced programmer tackle this project, click “get help” to see a walkthrough video.